pyproject.toml
[project]
name = "two-coins"
version = "0.0.1"
description = "Budgeting django app"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.3",
    "django-filter>=25.2",
    "python-dotenv>=1.2.2",
]
